requirements.txt
azure-ai-textanalytics
botbuilder-core
botbuilder-dialogs
botbuilder-schema
flask
